# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: External Trading Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List signal history
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: finalSignal
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Signal list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
...
  /signals/generate:
    post:
      operationId: generateSignal
      tags: [signals]
      summary: Manually trigger signal generation for a symbol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalRequest"
      responses:
        "200":
          description: Generated signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
...
  /webhook/tradingview:
    post:
      operationId: receiveTradingViewAlert
      tags: [webhook]
      summary: Receive alert from TradingView
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradingViewAlert"
      responses:
        "200":
          description: Alert received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookAck"
  /webhook/send:
    post:
      operationId: sendSignalToTradingView
      tags: [webhook]
      summary: Manually send a signal to TradingView
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendSignalInput"
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookAck"
  /webhook/logs:
    get:
      operationId: listWebhookLogs
      tags: [webhook]
      summary: Get webhook event log history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: direction
          in: query
          schema:
            type: string
            enum: [inbound, outbound]
      responses:
        "200":
          description: Webhook log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookLog"
...
components:
  schemas:
...
    Signal:
      type: object
      required: [id, symbol, finalSignal, confidence, macroSignal, orderbookSignal, earningsSignal, technicalSignal, createdAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        finalSignal:
          type: string
          enum: [BUY, SELL, NONE]
        confidence:
          type: number
        macroSignal:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        orderbookSignal:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        earningsSignal:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        technicalSignal:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        score:
          type: ["number", "null"]
        sentToTradingView:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    SignalRequest:
      type: object
      required: [symbol]
      properties:
        symbol:
          type: string
        sendToTradingView:
          type: boolean
          default: false
...
    TradingViewAlert:
      type: object
      required: [symbol, action]
      properties:
        symbol:
          type: string
        action:
          type: string
        price:
          type: ["number", "null"]
        message:
          type: ["string", "null"]
    WebhookAck:
      type: object
      required: [status]
      properties:
        status:
          type: string
        message:
          type: ["string", "null"]
    SendSignalInput:
      type: object
      required: [symbol, finalSignal]
      properties:
        symbol:
          type: string
        finalSignal:
          type: string
          enum: [BUY, SELL, NONE]
        confidence:
          type: number
          default: 0
    WebhookLog:
      type: object
      required: [id, direction, payload, status, createdAt]
      properties:
        id:
          type: integer
        direction:
          type: string
          enum: [inbound, outbound]
        payload:
          type: string
          description: JSON-serialized webhook payload
        status:
          type: string
          enum: [success, error, received]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...